+++ rollup.config.mjs
@@ -6,7 +6,7 @@
 import { nodeResolve } from "@rollup/plugin-node-resolve";
 
 export default () => ({
-	input: "src/module/condition-lab-triggler.js",
+	input: "src/module/ironsworn-impacts.js",
 	output: {
 		dir: "dist/module",
 		format: "es",