+++ src/module/condition-lab-triggler.js
@@ -32,7 +32,7 @@ Hooks.on("i18nInit", () => {
 	Sidekick.loadTemplates();
 
 	// Keybinds
-	game.keybindings.register("condition-lab-triggler", "openConditionLab", {
+	game.keybindings.register("ironsworn-impacts", "openConditionLab", {
 		name: "CLT.KEYBINDINGS.openConditionLab.name",
 		onDown: () => {
 			new ConditionLab().render(true);
@@ -40,7 +40,7 @@ Hooks.on("i18nInit", () => {
 		restricted: true,
 		precedence: CONST.KEYBINDING_PRECEDENCE.NORMAL
 	});
-	game.keybindings.register("condition-lab-triggler", "openTriggler", {
+	game.keybindings.register("ironsworn-impacts", "openTriggler", {
 		name: "CLT.KEYBINDINGS.openTriggler.name",
 		onDown: () => {
 			new TrigglerForm().render(true);
@@ -70,10 +70,10 @@ Hooks.on("i18nInit", () => {
 			}
 		};
 		libWrapper.register(
-			"condition-lab-triggler",
+			"ironsworn-impacts",
 			"Token.prototype._refreshEffects",
 			function () {
-				const effectSize = game.settings.get("condition-lab-triggler", "effectSize");
+				const effectSize = game.settings.get("ironsworn-impacts", "effectSize");
 				// Use the default values if no setting found
 				const { multiplier = 2, divisor = 5 } = effectSizes[effectSize];
 
@@ -130,10 +130,10 @@ Hooks.on("ready", async () => {
 		});
 	game.clt.CoreSpecialStatusEffects = foundry.utils.deepClone(CONFIG.specialStatusEffects);
 	game.clt.supported = false;
-	let defaultMaps = game.settings.get("condition-lab-triggler", "defaultConditionMaps");
-	let conditionMap = game.settings.get("condition-lab-triggler", "activeConditionMap");
+	let defaultMaps = game.settings.get("ironsworn-impacts", "defaultConditionMaps");
+	let conditionMap = game.settings.get("ironsworn-impacts", "activeConditionMap");
 
-	const mapType = game.settings.get("condition-lab-triggler", "conditionMapType");
+	const mapType = game.settings.get("ironsworn-impacts", "conditionMapType");
 
 	// If there's no defaultMaps or defaultMaps doesn't include game system, check storage then set appropriately
 	if (
@@ -145,12 +145,12 @@ Hooks.on("ready", async () => {
 		)
 	) {
 		defaultMaps = await EnhancedConditions._loadDefaultMaps();
-		game.settings.set("condition-lab-triggler", "defaultConditionMaps", defaultMaps);
+		game.settings.set("ironsworn-impacts", "defaultConditionMaps", defaultMaps);
 	}
 
 	// If map type is not set and a default map exists for the system, set maptype to default
 	if (!mapType && defaultMaps instanceof Object && Object.keys(defaultMaps).includes(game.system.id)) {
-		game.settings.set("condition-lab-triggler", "conditionMapType", "default");
+		game.settings.set("ironsworn-impacts", "conditionMapType", "default");
 	}
 
 	// If there's no condition map, get the default one
@@ -159,13 +159,13 @@ Hooks.on("ready", async () => {
 		conditionMap = EnhancedConditions.getDefaultMap(defaultMaps);
 
 		if (game.user.isGM && conditionMap.length) {
-			game.settings.set("condition-lab-triggler", "activeConditionMap", conditionMap);
+			game.settings.set("ironsworn-impacts", "activeConditionMap", conditionMap);
 		}
 	}
 
 	// If map type is not set, now set to default
 	if (!mapType && conditionMap.length) {
-		game.settings.set("condition-lab-triggler", "conditionMapType", "default");
+		game.settings.set("ironsworn-impacts", "conditionMapType", "default");
 	}
 
 	// Update status icons accordingly
@@ -173,7 +173,7 @@ Hooks.on("ready", async () => {
 		// CONFIG.statusEffects
 		// CONFIG.specialStatusEffects
 	}
-	// const specialStatusEffectMap = game.settings.get("condition-lab-triggler", "specialStatusEffectMapping");
+	// const specialStatusEffectMap = game.settings.get("ironsworn-impacts", "specialStatusEffectMapping");
 	if (conditionMap.length) EnhancedConditions._updateStatusEffects(conditionMap);
 	setInterval(EnhancedConditions.updateConditionTimestamps, 15000);
 
@@ -210,8 +210,8 @@ Hooks.on("deleteActiveEffect", (effect, options, userId) => {
 /* ------------------ Combat ------------------ */
 
 Hooks.on("updateCombat", (combat, update, options, userId) => {
-	const enableOutputCombat = game.settings.get("condition-lab-triggler", "conditionsOutputDuringCombat");
-	const outputChatSetting = game.settings.get("condition-lab-triggler", "conditionsOutputToChat");
+	const enableOutputCombat = game.settings.get("ironsworn-impacts", "conditionsOutputDuringCombat");
+	const outputChatSetting = game.settings.get("ironsworn-impacts", "conditionsOutputToChat");
 	const combatant = combat.combatant;
 
 	if (
@@ -243,7 +243,7 @@ Hooks.on("updateCombat", (combat, update, options, userId) => {
 
 /* -------------- Scene Controls -------------- */
 Hooks.on("getSceneControlButtons", function (hudButtons) {
-	if (game.user.isGM && game.settings.get("condition-lab-triggler", "sceneControls")) {
+	if (game.user.isGM && game.settings.get("ironsworn-impacts", "sceneControls")) {
 		const hud = $(hudButtons).find((val) => val.name === "token");
 		if (hud) {
 			hud.tools.push({
@@ -283,7 +283,7 @@ Hooks.on("renderSceneControls", (app, html, data) => {
 /* ------------------- Misc ------------------- */
 
 Hooks.on("renderSettingsConfig", (app, html, data, ...others) => {
-	const trigglerMenu = $(html).find("button[data-key=\"condition-lab-triggler.trigglerMenu\"]")[0];
+	const trigglerMenu = $(html).find("button[data-key=\"ironsworn-impacts.trigglerMenu\"]")[0];
 	if (trigglerMenu) {
 		const exclamationMark = trigglerMenu.children[0];
 		exclamationMark.style.margin = "0 -6px";
@@ -299,11 +299,11 @@ Hooks.on("renderSettingsConfig", (app, html, data, ...others) => {
 Hooks.on("renderMacroConfig", (app, html, data) => {
 	const typeSelect = $(html).find("select[name='type']");
 	const typeSelectDiv = typeSelect.closest("div");
-	const flag = app.object.getFlag("condition-lab-triggler", "macroTrigger");
-	const triggers = game.settings.get("condition-lab-triggler", "storedTriggers");
+	const flag = app.object.getFlag("ironsworn-impacts", "macroTrigger");
+	const triggers = game.settings.get("ironsworn-impacts", "storedTriggers");
 
 	const select = foundry.applications.fields.createSelectInput({
-		name: "flags.condition-lab-triggler.macroTrigger",
+		name: "flags.ironsworn-impacts.macroTrigger",
 		options: triggers,
 		value: flag,
 		blank: "CLT.ENHANCED_CONDITIONS.MacroConfig.NoTriggerSet",