+++ src/module/sidekick.js
@@ -127,9 +127,9 @@ export class Sidekick {
 	 */
 	static async loadTemplates() {
 		const templates = [
-			"modules/condition-lab-triggler/templates/partials/chat-card-condition-list.hbs",
-			"modules/condition-lab-triggler/templates/partials/condition-lab-row.hbs",
-			"modules/condition-lab-triggler/templates/partials/triggler-icon.hbs"
+			"modules/ironsworn-impacts/templates/partials/chat-card-condition-list.hbs",
+			"modules/ironsworn-impacts/templates/partials/condition-lab-row.hbs",
+			"modules/ironsworn-impacts/templates/partials/triggler-icon.hbs"
 		];
 		await loadTemplates(templates);
 	}